backend/requirements.txt
pymysql
sqlalchemy
PyJWT
bcrypt
reportlab
qrcode
pypdf2
flask 
flask_cors
user_agents
cryptography